The file scrapping.ipynb contains the code to scrap content from NPO webpage by category. For each item we get: name, category (programma, film, documentaire or serie), description (if available) and the images. At the end, we will get eight different csv files: documentaires, films, series and programmas, that contains the information found in the NPO platform, and documentaires_complete, films_complete, programmas_complete and films_complete that contains more information that was found in the movie meter webpage.
Then, we complete the data by scrapping movie meter webpage. In this case, we look for the show in the webpage and then, we get the description (if it was not found in NPO), images (if we didn't find any in NPO) and tags (komedie, animatie, familie, etc). At the end, we get the Final_NPO_Data, that contains the items that will be used in the project.